Redo conversion + tweaks
Use ElementTree to read and write XML instead of manually via text file (what was I thinking)
Move list of objects to the main file, and pass it to the GUI
Fix GUI loading prefs and selecting objects
Organization tweaks

+++ rmj2gms.py
@@ -1,48 +1,53 @@
-# entry point, input sanitization
+# entry point, input sanitization, general controller stuff
 
-import convert, gui, localize
-import os, sys
+import os
+import gui, convert, localize
 
 def loc(key):
     return localize.loc(key)
 
-def save_prefs(template, project, objects):
-    with open('prefs', 'w') as f:
-        f.write('template|%s\n' % template)
-        f.write('project|%s\n' % project)
-        for key, value in sorted(objects.items()):
-            f.write('%s|%s|%s\n' % (key, value[0], value[1]))
+# object_inputs is a map: {object_name: (gm_object_name, enabled)}
+def submitted(project_path, template_room_name, map_path, object_inputs):
 
-def convert_pressed(rmj, template, project, objects):
-    if rmj == '' or template == '' or project == '':
+    # input sanitization
+    if project_path == '' or template_room_name == '' or map_path == '':
         return loc('error_top_field_empty')
-    for key, val in objects.items():
-        text = val[0]
-        enabled = val[1]
+    if not os.path.exists(project_path):
+        # TODO localize
+        return 'project does not exist'
+    if not os.path.exists(map_path):
+        # TODO localize
+        return 'map does not exist'
+    template_room_path = os.path.join(os.path.split(project_path)[0], 'rooms', template_room_name+'.room.gmx')
+    if not os.path.exists(template_room_path):
+        # TODO localize
+        return 'template room does not exist'
+    for objectname, enabled in object_inputs.values():
         if enabled:
-            if text == '':
+            if objectname == '':
                 return loc('error_object_no_name')
-            else:
-                fn = os.path.join(os.path.split(project)[0], 'objects', text + '.object.gmx')
-                if not os.path.exists(fn):
-                    return loc('error_nonexistent_object') % text
-    roomname, fn = convert.get_room_names(rmj, project)
-    try:
-        rmj_to_gm = {}
-        for key, val in objects.items():
-            if val[1] == 1:
-                rmj_to_gm[key] = val[0]
-        entities = convert.get_entities_from_rmj(rmj, rmj_to_gm)
-        convert.write_room(entities, roomname, project, template)
-        convert.add_room_to_project(roomname, project)
-        save_prefs(template=template, project=project, objects=objects)
-        return loc('convert_successful') + ' ' + roomname
-    except:
-        info = sys.exc_info()
-        with open('errorlog.txt', 'w') as f:
-            f.write('Last error:\n\n%s\n%s\n%s' % info)
-        return loc('error_exception')
+            fn = os.path.join(os.path.split(project_path)[0], 'objects', objectname + '.object.gmx')
+            if not os.path.exists(fn):
+                return loc('error_nonexistent_object') % objectname
 
+    # build dict of object names that were enabled
+    chosen_object_names = {}
+    for object_name, (gm_object_name, enabled) in object_inputs.items():
+        if enabled:
+            chosen_object_names[object_name] = gm_object_name
+
+    output_room_name = convert.convert(project_path, template_room_path, map_path, chosen_object_names)
+
+    # save preferences
+    with open('prefs', 'w') as f:
+        f.write('project|%s\n' % project_path)
+        f.write('template|%s\n' % template_room_name)
+        for object_name, (gm_object_name, enabled) in sorted(object_inputs.items()):
+            f.write('%s|%s|%i\n' % (object_name, gm_object_name, enabled))
+
+    return loc('convert_successful') + ' ' + output_room_name
+
+# prompt and load language
 if os.path.exists('lang'):
     with open('lang', 'r') as f:
         language = list(f)[0]
@@ -54,4 +59,5 @@ else:
     localize.load(language)
     gui.show_instructions()
 
-gui.run(submit_func=convert_pressed)
+# kick off the gui
+gui.run(submit_func=submitted)
